Moved code from wrong adagrams.rb into lib/adagrams.rb
